Move ProjectSearch to FullTextSearch
