+++ pages/_app.tsx
@@ -39,7 +39,7 @@ import "devtools/client/debugger/src/components/Editor/SearchBar.css";
 import "devtools/client/debugger/src/components/Editor/Tabs.css";
 import "devtools/client/debugger/src/components/PrimaryPanes/Outline.css";
 import "devtools/client/debugger/src/components/PrimaryPanes/Sources.css";
-import "devtools/client/debugger/src/components/ProjectSearch.css";
+import "devtools/client/debugger/src/components/FullTextSearch/FullTextSearch.css";
 import "devtools/client/debugger/src/components/QuickOpenModal.css";
 import "devtools/client/debugger/src/components/SecondaryPanes/Breakpoints/Breakpoint.css";
 import "devtools/client/debugger/src/components/SecondaryPanes/Breakpoints/BreakpointNavigation.css";
