+++ src/devtools/client/debugger/src/components/FullTextSearch/FullTextSearch.css
