+++ src/ui/components/SidePanel.tsx
@@ -6,7 +6,7 @@ import Transcript from "ui/components/Transcript";
 import Events from "ui/components/Events";
 import ReplayInfo from "./Events/ReplayInfo";
 import PrimaryPanes from "devtools/client/debugger/src/components/PrimaryPanes";
-const ProjectSearch = require("devtools/client/debugger/src/components/ProjectSearch").default;
+const FullTextSearch = require("devtools/client/debugger/src/components/FullTextSearch").default;
 const SecondaryPanes = require("devtools/client/debugger/src/components/SecondaryPanes").default;
 const Accordion = require("devtools/client/debugger/src/components/shared/Accordion").default;
 
@@ -24,7 +24,7 @@ function SidePanel({ selectedPrimaryPanel }: SidePanelProps) {
   } else if (selectedPrimaryPanel === "comments") {
     sidepanel = <Transcript />;
   } else if (selectedPrimaryPanel === "search") {
-    sidepanel = <ProjectSearch />;
+    sidepanel = <FullTextSearch />;
   } else if (selectedPrimaryPanel === "events") {
     sidepanel = (
       <Accordion
